backspace implemented
Backspace works in all the basic cases.
The difficulty arises when backspace is used for chaining operations.
The current implementation doesn't allow to implement backspace for
chaining operations. I am intending to branch from main and experiment with
another architecture: instead of using two variables for keeping number
values I'll be trying to use an array of variables to hopefully be able
implement chaining and correspnding backspacing...