additional functionality of '-'
Problem: before one couldn't do the following: 6*-2 = -12
NOw '-' operation can also serve as a sign settler.
In addition to this functionality there will be '+/-' button for
complementary functionality. For example with current implementation
of '-' it is not possible to change the sign of the resulting
number (after hitting '=').
Also I've renamed the variable isNew to isResult. It makes more sense
for the context